+++ cmd/influxd/inspect/verify_seriesfile.go
@@ -1,6 +1,7 @@
-package seriesfile
+package inspect
 
 import (
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -9,9 +10,93 @@ import (
 	"sort"
 	"sync"
 
+	"github.com/influxdata/influxdb/v2/logger"
 	"github.com/influxdata/influxdb/v2/tsdb"
+	"github.com/spf13/cobra"
 	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
 )
+
+type args struct {
+	dir        string
+	db         string
+	seriesFile string
+	verbose    bool
+	concurrent int
+}
+
+func NewVerifySeriesfileCommand() *cobra.Command {
+	var arguments args
+	cmd := &cobra.Command{
+		Use:   "verify-seriesfile",
+		Short: "Verifies the integrity of series files.",
+		Args:  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			cmd.SetOut(os.Stdout)
+
+			config := logger.NewConfig()
+			config.Level = zapcore.WarnLevel
+			if arguments.verbose {
+				config.Level = zapcore.InfoLevel
+			}
+			log, err := config.New(os.Stderr)
+			if err != nil {
+				return err
+			}
+
+			v := NewVerify()
+			v.Logger = log
+			v.Concurrent = arguments.concurrent
+
+			var db string
+			if arguments.seriesFile != "" {
+				db = arguments.seriesFile
+			} else if arguments.db != "" {
+				db = filepath.Join(arguments.dir, arguments.db, "_series")
+			}
+			if db != "" {
+				_, err := v.VerifySeriesFile(db)
+				return err
+			}
+
+			dbs, err := ioutil.ReadDir(arguments.dir)
+			if err != nil {
+				return err
+			}
+
+			var hasError bool
+			for _, db := range dbs {
+				if !db.IsDir() {
+					continue
+				}
+				filePath := filepath.Join(arguments.dir, db.Name(), "_series")
+				if _, err := v.VerifySeriesFile(filePath); err != nil {
+					v.Logger.Error("Failed to verify series file",
+						zap.String("filename", filePath),
+						zap.Error(err))
+					hasError = true
+				}
+			}
+			if hasError {
+				return errors.New("some files failed verification, see logs for details")
+			}
+			return nil
+		},
+	}
+
+	cmd.Flags().StringVar(&arguments.dir, "dir", filepath.Join(os.Getenv("HOME"), ".influxdbv2", "engine", "data"),
+		"Data Directory.")
+	cmd.Flags().StringVar(&arguments.db, "db", "",
+		"Only use this database inside of the data directory.")
+	cmd.Flags().StringVar(&arguments.seriesFile, "series-file", "",
+		"Path to a series file. This overrides --db and --dir.")
+	cmd.Flags().BoolVar(&arguments.verbose, "v", false,
+		"Verbose output.")
+	cmd.Flags().IntVar(&arguments.concurrent, "c", runtime.GOMAXPROCS(0),
+		"How many concurrent workers to run.")
+
+	return cmd
+}
 
 // verifyResult contains the result of a Verify... call
 type verifyResult struct {